% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fatty_acid_metabolism}
\alias{fatty_acid_metabolism}
\title{Mouse gene symbols related to fatty acid metabolism pathway}
\format{
A vector of character with 158 elements.
}
\usage{
fatty_acid_metabolism
}
\description{
A vector of character containing 158 gene symbols related with the fatty acid metabolism pathway in Mouse.
}
\examples{
data(fatty_acid_metabolism)

}
\keyword{datasets}
