% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways}
\alias{pathways}
\title{Pathways with their associated genes for Mouse}
\format{
A data frame with 7393 observations of 2 variables.
\describe{
  \item{gs_name}{50 pathways from Mouse.}
  \item{gene_symbol}{4398 gene symbols from Mouse}
}
}
\usage{
pathways
}
\description{
A data frame with pathways in one column (gs_name) and their associated genes (gene_symbol).
}
\examples{
data(pathways)

}
\keyword{datasets}
