% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ranked.list}
\alias{ranked.list}
\title{Ranked list of genes from gene expression experiment.}
\format{
A vector of character with 24421 elements.
}
\usage{
ranked.list
}
\description{
A vector of character ranked by logFC of each gene from starving and normal condition in Mouse experiment.
}
\examples{
data(ranked.list)

}
\keyword{datasets}
