% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_set_gaf}
\alias{read_set_gaf}
\title{Read .gaf file}
\usage{
read_set_gaf(gaf, aspect)
}
\arguments{
\item{gaf}{a gaf file}

\item{aspect}{one of the GO aspect :"C" : Cellular component "P" : Biological process "M" : Molecular function}
}
\value{
a data frame containing sets with their corresponding gene symbols. Columns: gs_name, gene_symbol.
}
\description{
This function reads .gaf file and format it to be used by analysis functions for a given GO aspect.
}
\details{
This function also maps GO ids to their corresponding terms with the AnnotationDbi package via mapIds function.
}
\examples{
data(gaf_file)
read_set_gaf(gaf_file, "C")
read_set_gaf(gaf_file, "M")
read_set_gaf(gaf_file, "P")
}
